[issue-51] Remove inheritance in HiveMetaStoreClientFactory
