+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ### Added
 * Regex support in `federated-meta-stores.mapped-databases`. [#59](https://github.com/HotelsDotCom/waggle-dance/issues/59).
 ### Changed
-* Replace SSH support with _hcommon-ssh_ library. [#51](https://github.com/HotelsDotCom/waggle-dance/issues/51).
+* Replace SSH support with [hcommon-ssh](https://github.com/HotelsDotCom/hcommon-ssh) library. [#51](https://github.com/HotelsDotCom/waggle-dance/issues/51).
 
 ## [2.3.1] 2018-08-09
 ### Added
