+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/client/MetaStoreClientFactory.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2017 Expedia Inc.
+ * Copyright (C) 2016-2018 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,101 +15,10 @@
  */
 package com.hotels.bdp.waggledance.client;
 
-import java.lang.reflect.InvocationHandler;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import java.lang.reflect.Proxy;
-
 import org.apache.hadoop.hive.conf.HiveConf;
-import org.apache.thrift.transport.TTransportException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class MetaStoreClientFactory {
+public interface MetaStoreClientFactory {
 
-  static final Class<?>[] INTERFACES = new Class<?>[] { CloseableThriftHiveMetastoreIface.class };
-
-  private static class ReconnectingMetastoreClientInvocationHandler implements InvocationHandler {
-    private static final Logger LOG = LoggerFactory.getLogger(ReconnectingMetastoreClientInvocationHandler.class);
-
-    private final ThriftMetastoreClient base;
-    private final String name;
-    private final int maxRetries;
-
-    private ReconnectingMetastoreClientInvocationHandler(HiveConf hiveConf, String name, int maxRetries) {
-      this.name = name;
-      this.maxRetries = maxRetries;
-      base = new ThriftMetastoreClient(hiveConf);
-    }
-
-    @Override
-    public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
-      int attempt = 0;
-      // close() and isOpen() methods delegate to base HiveMetastoreClient
-      switch (method.getName()) {
-      case "isOpen":
-        try {
-          reconnectIfDisconnected();
-          return base.isOpen();
-        } catch (Exception e) {
-          LOG.debug("Error re-opening client at isOpen(): {}", e.getMessage());
-          return false;
-        }
-      case "close":
-        if (base != null) {
-          base.close();
-        }
-        return null;
-      default:
-        reconnectIfDisconnected();
-        do {
-          try {
-            return method.invoke(base.getClient(), args);
-          } catch (InvocationTargetException e) {
-            Throwable realException = e.getTargetException();
-            if (TTransportException.class.isAssignableFrom(realException.getClass())) {
-              if (attempt < maxRetries && shouldRetry(method)) {
-                LOG.debug("TTransportException captured in client {}. Reconnecting... ", name);
-                base.reconnect();
-                continue;
-              }
-              throw new MetastoreUnavailableException("Client " + name + " is not available", realException);
-            }
-            throw realException;
-          }
-        } while (++attempt <= maxRetries);
-        break;
-      }
-      throw new RuntimeException("Unreachable code");
-
-    }
-
-    private boolean shouldRetry(Method method) {
-      switch (method.getName()) {
-      case "shutdown":
-        return false;
-      default:
-        return true;
-      }
-    }
-
-    private void reconnectIfDisconnected() {
-      try {
-        if (!base.isOpen()) {
-          base.reconnect();
-        }
-      } catch (Exception e) {
-        throw new MetastoreUnavailableException("Client " + name + " is not available", e);
-      }
-    }
-
-  }
-
-  public CloseableThriftHiveMetastoreIface newInstance(HiveConf hiveConf, String name, int reconnectionRetries) {
-    ReconnectingMetastoreClientInvocationHandler reconnectingHandler = new ReconnectingMetastoreClientInvocationHandler(
-        hiveConf, name, reconnectionRetries);
-    return (CloseableThriftHiveMetastoreIface) Proxy.newProxyInstance(getClass().getClassLoader(), INTERFACES,
-        reconnectingHandler);
-  }
+  CloseableThriftHiveMetastoreIface newInstance(HiveConf hiveConf, String name, int reconnectionRetries);
 
 }
