+++ waggle-dance-core/src/test/java/com/hotels/bdp/waggledance/client/CloseableThriftHiveMetastoreIfaceClientFactoryTest.java
@@ -42,7 +42,7 @@ public class CloseableThriftHiveMetastoreIfaceClientFactoryTest {
   private static final String THRIFT_URI = "thrift://host:port";
 
   private CloseableThriftHiveMetastoreIfaceClientFactory factory;
-  private @Mock MetaStoreClientFactory metaStoreClientFactory;
+  private @Mock DefaultMetaStoreClientFactory metaStoreClientFactory;
 
   @Before
   public void setUp() {
